Extract koans DSL to its own file/module
